Changed Float[] to Double[] in ZernikeMap

+++ server/src/main/java/uk/ac/cam/cl/echo/extrusionfinder/server/parts/ZernikeMap.java
@@ -14,10 +14,10 @@ import java.util.Map;
  */
 public class ZernikeMap implements DatabaseItem {
     private final String _id;
-    private final Map<String, Float[]> zernikeMap;
+    private final Map<String, Double[]> zernikeMap;
 
     @JsonCreator
-    public ZernikeMap(@JsonProperty("_id") String _id, @JsonProperty("zernikeMap") Map<String, Float[]> map) {
+    public ZernikeMap(@JsonProperty("_id") String _id, @JsonProperty("zernikeMap") Map<String, Double[]> map) {
         this._id = _id;
         this.zernikeMap = map;
     }
@@ -25,7 +25,7 @@ public class ZernikeMap implements DatabaseItem {
     /**
      * @param map Map from part identifiers to Zernike Moments
      */
-    public ZernikeMap(Map<String, Float[]> map) {
+    public ZernikeMap(Map<String, Double[]> map) {
         this._id = Configuration.ZERNIKE_MAP_ID;
         this.zernikeMap = map;
     }
@@ -38,7 +38,7 @@ public class ZernikeMap implements DatabaseItem {
     /**
      * @return Map from part identifiers to Zernike Moments
      */
-    public Map<String, Float[]> getZernikeMap() {
+    public Map<String, Double[]> getZernikeMap() {
         return zernikeMap;
     }
 }
